+++ lib/middleman-livereload.rb
@@ -2,6 +2,6 @@ require "middleman-core"
 require "middleman-livereload/version"
   
 ::Middleman::Extensions.register(:livereload) do
-  require "middleman-livereload/extension_3_1"
+  require "middleman-livereload/extension"
   ::Middleman::LiveReloadExtension
 end
