+++ lib/middleman-livereload/extension.rb
